+++ sdk/typespec/typespec_client_core/src/http/request/mod.rs
@@ -155,6 +155,11 @@ impl Request {
         &self.method
     }
 
+    /// Sets the request HTTP method.
+    pub fn set_method(&mut self, method: Method) {
+        self.method = method;
+    }
+
     /// Inserts zero or more headers from a type that implements [`AsHeaders`].
     pub fn insert_headers<T: AsHeaders>(&mut self, headers: &T) -> Result<(), T::Error> {
         for (name, value) in headers.as_headers()? {