+++ sdk/core/azure_core/CHANGELOG.md
@@ -7,10 +7,12 @@
 - Added `get_async_runtime()` and `set_async_runtime()` to allow customers to replace the asynchronous runtime used by the Azure SDK.
 - Added `PageIterator::continuation_token()` and `PageIterator::with_continuation_token()` to support reconstructing a `PageIterator` in another process or on another machine to continue paging.
 - Added `Poller<T>` for long-running operations (LROs).
+- Added `Request::set_method()` to allow changing the HTTP method of a request.
 - Added `StatusMonitor` for long-running operations.
 
 ### Breaking Changes
 
+- Added `http::PollerOptions` parameter to `http::poller::get_retry_after`.
 - Implemented `FromStr` where `FromStr::Err = Infallible` for `PollerStatus` instead of `From<&str>`.
 - Minimum supported Rust version (MSRV) is now 1.85.
 - `azure_core::http::Pipeline::new` now takes an `azure_core::http::ClientOptions` which is defined in `azure_core`, but convertible to `typespec_client_core::http::ClientOptions`.